Fixed missing MANIFEST and README description
The PyPI site requires plain text for the description, or valid RST but
the description extractor wasn't pulling out the RST footers containing
the link targets. Simpler just to remove them. The missing MANIFEST was
causing PyPI installs to fail as the setup.py calls utils.py which was
missing from the package.